neuralNetwork/requirements.txt
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
torch==2.2.1+cu121
pymongo==4.6.2
python-dotenv==1.0.1
